Se refactoriza la estructura de paquetes para mejorar la claridad y la separación de concerns, siguiendo las mejores prácticas de arquitectura.
- Se mueve la carpeta `Auth` y su contenido (DTOs de autenticación) al paquete principal de `DTOs`.
- Se centralizan todas las clases de DTOs (Data Transfer Objects) bajo un único paquete `com.Petcare.Petcare.DTOs`.

Este cambio distingue claramente las entidades de persistencia JPA (que permanecen en `Models`) de los objetos utilizados para la transferencia de datos en la API, haciendo el código más fácil de mantener y entender.
