+++ src/main/java/com/Petcare/Petcare/Configurations/Auth/AuthService.java
@@ -1,9 +1,9 @@
 package com.Petcare.Petcare.Configurations.Auth;
 
 import com.Petcare.Petcare.Configurations.Security.Jwt.JwtService;
-import com.Petcare.Petcare.Models.Auth.Request.LoginRequest;
-import com.Petcare.Petcare.Models.Auth.Request.RegisterRequest;
-import com.Petcare.Petcare.Models.Auth.Respone.AuthResponse;
+import com.Petcare.Petcare.DTOs.Auth.Request.LoginRequest;
+import com.Petcare.Petcare.DTOs.Auth.Request.RegisterRequest;
+import com.Petcare.Petcare.DTOs.Auth.Respone.AuthResponse;
 import com.Petcare.Petcare.Models.User.Role;
 import com.Petcare.Petcare.Models.User.User;
 import com.Petcare.Petcare.Repositories.UserRepository;
